+++ create-graphs.py
@@ -1,7 +1,5 @@
 from models.PyCryptoBot import PyCryptoBot
 from models.Trading import TechnicalAnalysis
-from models.Binance import AuthAPI as BAuthAPI, PublicAPI as BPublicAPI
-from models.CoinbasePro import AuthAPI as CBAuthAPI, PublicAPI as CBPublicAPI
 from views.TradingGraphs import TradingGraphs
 
 #app = PyCryptoBot()
