+++ models/TradingAccount.py
@@ -1,13 +1,12 @@
 """Live or test trading account"""
 
-import sys
 import numpy as np
 import pandas as pd
-import json, math, re, requests, sys
-from datetime import datetime
+import re, requests
 from binance.client import Client
-from models.Binance import AuthAPI as BAuthAPI, PublicAPI as BPublicAPI
-from models.CoinbasePro import AuthAPI as CBAuthAPI, PublicAPI as CBPublicAPI
+from models.exchange.binance import AuthAPI as BAuthAPI, PublicAPI as BPublicAPI
+from models.exchange.coinbase_pro import AuthAPI as CBAuthAPI
+
 
 class TradingAccount():
     def __init__(self, app=None):
