+++ tests/test_Model_CoinbasePro.py
@@ -3,7 +3,7 @@ from datetime import datetime
 
 sys.path.append('.')
 # pylint: disable=import-error
-from models.CoinbasePro import AuthAPI, PublicAPI
+from models.exchange.coinbase_pro import AuthAPI, PublicAPI
 
 VALID_ORDER_MARKET = 'BTC-GBP'
